requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
numpy==1.24.2
pandas==1.5.3
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
pillow==9.4.0
requests==2.28.2
gunicorn==20.1.0
